requirements.txt
# Flask and web
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.1

# Document processing
PyPDF2>=3.0.1
python-pptx>=0.6.23
python-docx>=1.1.0
striprtf>=0.0.26

# AI - Mistral
mistralai>=1.0.0

# Utilities
python-dotenv>=1.0.0
